Datatable compatibility
https: //github.com/callumok2004/gmod-dt-test/commit/da97f95eab0e362bf7fbf206a08751a774956bc1 for the diff check (made this MUCH easier)

+++ Game.Shared/BaseEntityShared.cs
@@ -147,7 +147,7 @@ public partial class
 		int minTick = TICK_NEVER_THINK;
 
 		if (NextThinkTick > 0)
-			minTick = NextThinkTick;
+			minTick = (int)NextThinkTick;
 
 		for (int i = 0; i < ThinkFunctions.Count; i++) {
 			int next = (int)ThinkFunctions[i].NextThinkTick;
